Summary of LR models performance
This is the results I got from LR. First I tried the data processing that we used for the other models (RF and GB) and I wasn't able to obtain a very good performance at all (Accuracy= 0.37). Also I was only using 5 % of the training data and the model was slow to train.

I decided to explore some kernels in Kaggle, and was able to find some examples on how to process the data in a convenient way for LR. The 2nd model predicts much better (Accuracy= 0.84) and is a lot faster too. The downside to this model is the number of False Negatives which can be a problem for our purposes.